refactor: move backend-independent canvas code out of rendering.c
